Improved old directory deletion handling

+++ driver.py
@@ -15,14 +15,17 @@ def main():
     test_executable = "test_persistence"
 
     # === Delete old data, model and plots ===
-    shutil.rmtree("data") # Delete old data folder along with any data
-    os.mkdir("data") # Create new empty data folder
+    if os.path.exists("data/"):
+        shutil.rmtree("data") # Delete old data folder along with any data
+        os.mkdir("data") # Create new empty data folder
 
-    shutil.rmtree("models") # Delete old models folder along with any models
-    os.mkdir("models") # Create new empty models folder
+    if os.path.exists("models/"):
+        shutil.rmtree("models") # Delete old models folder along with any models
+        os.mkdir("models") # Create new empty models folder
 
-    shutil.rmtree("plots") # Delete old plots folder along with any plots
-    os.mkdir("plots") # Create new empty plots folder
+    if os.path.exists("plots/"):
+        shutil.rmtree("plots") # Delete old plots folder along with any plots
+        os.mkdir("plots") # Create new empty plots folder
 
     # === Compile and run my_echolocation.c to generate new latencies ===
     print(f"\nCompiling '{latency_script}'...")
